requirements.txt
py-clob-client>=0.20.0
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0
eth-account>=0.11.0
